fix: general problems with dbconfig, and added UseCases as instructed before
