+++ back/src/index.ts
@@ -1,33 +1,31 @@
-import { ApolloServer } from 'apollo-server-express';
-import express from 'express';
-import { GraphQLError } from 'graphql';
-import { env } from 'process';
-import { dbConfig } from './db/dbconfig';
-import { resolvers } from './resolvers';
-import { typeDefs } from './schema';
-require("dotenv").config();
+import { ApolloServer } from "apollo-server-express"
+import express from "express"
+import { GraphQLError } from "graphql"
+import { env } from "process"
+import { dbConfig } from "./db/dbconfig"
+import { resolvers } from "./resolvers"
+import { typeDefs } from "./schema"
+require("dotenv").config()
 
 const app = express()
 
 const apolloServer = new ApolloServer({
-    typeDefs,
-    formatError: (error: GraphQLError) => {
-            return {
-                name: error.name,
-                message: error.message,
-            };
-    },
-    resolvers
+  typeDefs,
+  formatError: (error: GraphQLError) => {
+    return {
+      name: error.name,
+      message: error.message
+    }
+  },
+  resolvers
 })
 
-;(async function() {
-    await apolloServer.start()
-    apolloServer.applyMiddleware({app, path: "/graphql"})
-    await dbConfig()
+;(async function () {
+  await apolloServer.start()
+  apolloServer.applyMiddleware({ app, path: "/graphql" })
+  await dbConfig()
 
-    // antes do listen o dbconfig
-    app.listen(env.PORT || 3000, () => {
-        console.log(`http://${env.HOST}:%d/graphql`, env.PORT || 3000);
-        
-    })
+  app.listen(env.PORT || 3000, () => {
+    console.log(`http://${env.HOST}:%d/graphql`, env.PORT || 3000)
+  })
 })()
